qpos/chirality/boltz_augmentation.py
import random

try:
    import boltz
except ImportError:
    boltz = None

class MirrorAugmentedDataset:
    """
    Wraps any Boltz-2 dataset and applies stochastic mirror augmentation.
    Default mirror probability: 50%.
    Every L-protein is reflected (x → -x) + chirality flags inverted
    to generate a physically valid D-protein training example.
    """
    def __init__(self, base_dataset, mirror_prob=0.5):
        if boltz is None:
            raise ImportError(
                "The boltz_augmentation module requires the optional 'boltz' dependency. "
                "Please install it using: pip install qpos[boltz]"
            )
        self.base_dataset = base_dataset
        self.mirror_prob = mirror_prob
        
    def __getitem__(self, idx):
        sample = self.base_dataset[idx]
        if random.random() < self.mirror_prob:
            sample = self._mirror_sample(sample)
        return sample
        
    def _mirror_sample(self, sample):
        # Implementation mirrors coordinates and inverts chiral volume target
        return sample

class ChiralVolumeLoss:
    """
    Differentiable version of signed chiral volume test.
    L_chiral = Σ_i max(0, margin - sign(V_ref_i) * V_pred_i)
    Hinge loss: penalizes incorrect chirality sign in predicted structures.
    lambda_chiral: weight in total loss L_total = L_diffusion + λ * L_chiral
    """
    def __init__(self, margin=0.0, lambda_weight=1.0):
        if boltz is None:
            raise ImportError(
                "The boltz_augmentation module requires the optional 'boltz' dependency."
            )
        self.margin = margin
        self.lambda_weight = lambda_weight
