requirements.txt
pennylane>=0.38
pennylane-lightning>=0.38
biopython>=1.83
numpy>=1.26
scipy>=1.12
matplotlib>=3.8
seaborn>=0.13
torch>=2.2
pandas>=2.2
networkx>=3.2
scikit-learn>=1.4
lightgbm>=4.0
xgboost>=2.0
fair-esm>=2.0.0
peft>=0.10.0
pytorch-lightning>=2.2.0
qiskit-aer>=0.13
requests>=2.31
tqdm>=4.66
pytest>=8.0
pytest-cov>=4.0
pytest-mock>=3.12.0
pyyaml>=6.0
